Refactor: MobileLink 모듈 분리 및 header 구조 수정
