+++ components/header/mobile-nav.tsx
@@ -1,43 +1,8 @@
-'use client';
-
-import Link, { LinkProps } from 'next/link';
-import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 
-import { cn } from '~/libs/utils';
-import { PropsWithRequiredChildren } from '~/types/utils';
+import { Sheet, SheetContent, SheetTrigger } from '~/components/sheet';
 
-import { Sheet, SheetContent, SheetTrigger } from './sheet';
-
-interface MobileLinkProps extends PropsWithRequiredChildren<LinkProps> {
-  href: string;
-  onOpenChange: (isOpen: boolean) => void;
-  className?: string;
-}
-
-const MobileLink = ({
-  href,
-  onOpenChange,
-  className,
-  children,
-  ...props
-}: MobileLinkProps) => {
-  const router = useRouter();
-
-  return (
-    <Link
-      href={href}
-      onClick={() => {
-        router.push(href.toString());
-        onOpenChange?.(false);
-      }}
-      className={cn(className)}
-      {...props}
-    >
-      {children}
-    </Link>
-  );
-};
+import MobileLink from './mobile-link';
 
 const MobileNav = () => {
   const [isOpen, setIsOpen] = useState(false);
